requirements.txt
tensorflow==2.10.0
keras==2.10.0
scikit-learn==1.0.2
flask==2.0.1
flask-restful==0.3.9
gunicorn==20.1.0
pillow==9.0.0
opencv-python==4.5.5.64
numpy==1.21.0
pandas==1.3.0
matplotlib==3.5.0
seaborn==0.11.2
shap==0.40.0
lime==0.2.0
pytest==6.2.5
